+++ app/src/main/java/com/codesoom/assignment/filters/AuthenticationErrorFilter.java
@@ -1,0 +1,28 @@
+package com.codesoom.assignment.filters;
+
+import com.codesoom.assignment.errors.InvalidTokenException;
+import org.springframework.http.HttpStatus;
+
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpFilter;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class AuthenticationErrorFilter extends HttpFilter {
+    @Override
+    protected void doFilter(HttpServletRequest request,
+                            HttpServletResponse response,
+                            FilterChain chain)
+            throws IOException, ServletException {
+       try{
+           chain.doFilter(request, response);
+
+       } catch (InvalidTokenException e) {
+           response.sendError(HttpStatus.UNAUTHORIZED.value());
+       }
+
+    }
+}
