+++ app/src/main/java/com/codesoom/assignment/filters/JwtAuthenticationFilter.java
@@ -1,6 +1,7 @@
 package com.codesoom.assignment.filters;
 
 import com.codesoom.assignment.application.AuthenticationService;
+import com.codesoom.assignment.errors.InvalidTokenException;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
@@ -25,22 +26,42 @@ public class JwtAuthenticationFilter extends BasicAuthenticationFilter {
                                     HttpServletResponse response,
                                     FilterChain chain)
             throws IOException, ServletException {
-        
+        if(filterWithPathAndMethod(request)) {
+            chain.doFilter(request, response);
+            return ;
+        }
+
         String authorization = request.getHeader("Authorization");
 
         if (authorization == null) {
-            response.sendError(HttpStatus.UNAUTHORIZED.value());
-            return;
+            throw new InvalidTokenException("");
         }
 
         String accessToken = authorization.substring("Bearer ".length());
-        Long userId = authenticationService.parseToken(accessToken);
 
+        Long userId = authenticationService.parseToken(accessToken);
         request.setAttribute("userId", userId);
 
         chain.doFilter(request, response);
 
     }
 
+    private boolean filterWithPathAndMethod(HttpServletRequest request) {
+        String path = request.getRequestURI();
+        if (!path.startsWith("/products")) {
+            return true;
+        }
+
+        String method = request.getMethod();
+        if (method.equals("GET")) {
+            return true;
+        }
+
+        if (method.equals("OPTIONS")) {
+            return true;
+        }
+
+        return false;
+    }
 
 }
